+++ MovieRental/ViewModels/CustomerViewModel.cs
@@ -5,11 +5,11 @@ namespace MovieRental
 {
 	public class CustomerViewModel
 	{
-        public  Dictionary<IPurchase, KeyValuePair<int, int>> RentalsWithPrices;
-		public int Total;
+        public  Dictionary<IPurchase, KeyValuePair<double, int>> RentalsWithPrices;
+        public double Total;
         public int Points;
 
-        public CustomerViewModel ( Dictionary<IPurchase, KeyValuePair<int, int>> rentalsWithPrices, int total, int points)
+        public CustomerViewModel ( Dictionary<IPurchase, KeyValuePair<double, int>> rentalsWithPrices, double total, int points)
 		{
 			this.Total = total;
 			this.RentalsWithPrices = rentalsWithPrices;
